Sparate public from internal classes.
