+++ lib/archethic_wallet_client.dart
@@ -2,9 +2,17 @@
 library archethic_wallet_client;
 
 export 'src/archethic_wallet_client_base.dart';
-export 'src/core/failures.dart';
-export 'src/core/request.dart';
-export 'src/core/result.dart';
-export 'src/core/subscription.dart';
-export 'src/model/app_account.dart';
-export 'src/request/requests.dart';
+export 'src/model/failures.dart';
+export 'src/model/request.dart';
+export 'src/model/session.dart';
+export 'src/model/subscription.dart';
+export 'src/request/account_sub.dart';
+export 'src/request/create_session.dart';
+export 'src/request/get_accounts.dart';
+export 'src/request/get_current_account.dart';
+export 'src/request/get_endpoint.dart';
+export 'src/request/get_services_from_keychain.dart';
+export 'src/request/keychain_derive_address.dart';
+export 'src/request/keychain_derive_keypair.dart';
+export 'src/request/send_transaction.dart';
+export 'src/request/sign_transactions.dart';
