+++ lib/src/core/result.dart
@@ -1,5 +1,5 @@
 /// SPDX-License-Identifier: AGPL-3.0-or-later
-import 'package:archethic_wallet_client/src/core/failures.dart';
+import 'package:archethic_wallet_client/src/model/failures.dart';
 import 'package:flutter/foundation.dart';
 
 extension FutureResult<ValueT, FailureT extends Exception>
