+++ lib/src/model/failures.dart
