+++ lib/src/model/failures.freezed.dart
