+++ lib/src/request/get_accounts.dart
@@ -1,9 +1,22 @@
 /// SPDX-License-Identifier: AGPL-3.0-or-later
-import 'package:archethic_wallet_client/src/model/app_account.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
 part 'get_accounts.freezed.dart';
 part 'get_accounts.g.dart';
+
+@freezed
+class AppAccount with _$AppAccount {
+  const factory AppAccount({
+    required String shortName,
+    required String serviceName,
+    required String genesisAddress,
+  }) = _AppAccount;
+
+  const AppAccount._();
+
+  factory AppAccount.fromJson(Map<String, dynamic> json) =>
+      _$AppAccountFromJson(json);
+}
 
 @freezed
 class GetAccountsResult with _$GetAccountsResult {
